+++ src/main/java/use_case/cat_management/change_cat_happiness/ChangeCatHappinessInteractor.java
@@ -43,17 +43,19 @@ public class ChangeCatHappinessInteractor implements ChangeCatHappinessInputBoun
     }
 
     int calculateHappinessPoints(int studySessionLength) {
-        if (studySessionLength <= 20) {
-            return Constants.POINTS_FOR_LESS_EQUAL_20;
+        final int happinessPoints;
+        if (studySessionLength <= Constants.MINUTES_20) {
+            happinessPoints = Constants.POINTS_FOR_LESS_EQUAL_20;
         }
-        else if (studySessionLength <= 40) {
-            return Constants.POINTS_FOR_BETWEEN_21_AND_40;
+        else if (studySessionLength <= Constants.MINUTES_40) {
+            happinessPoints = Constants.POINTS_FOR_BETWEEN_21_AND_40;
         }
-        else if (studySessionLength < 60) {
-            return Constants.POINTS_FOR_BETWEEN_41_AND_59;
+        else if (studySessionLength < Constants.MINUTES_60) {
+            happinessPoints = Constants.POINTS_FOR_BETWEEN_41_AND_59;
         }
         else {
-            return Constants.POINTS_FOR_60;
+            happinessPoints = Constants.POINTS_FOR_60;
         }
+        return happinessPoints;
     }
 }
